- Fix for compilation issues per Simon - Changed dependency on MPI to Lemon for those that use Lemon.   You can do MPI without Lemon, but not Lemon without MPI. - Added a MILC to ILDG conversion utility.   Discovered some bugs in the MILC checksum computation, not yet fixed - SaveILDG did not fully confrom to ILDG standards, it did not include   the ildg-format record. It does now - SaveILDG does not include a Scidac checksum, which is required by   loadILDG, this discrepancy needs fixing.

+++ src/L0/Core/Correlator/Correlator_printWithMomentum_full_Cstyle.cpp
@@ -1,12 +1,12 @@
 #include "Correlator.ih"
-
+#include <cstdio>
 namespace Core
 {
 
   template< >
   void Correlator< Dirac::Matrix >::printWithMomentum_full_Cstyle(FILE * out, int const * const momentum, std::string const& prefix) const
   {
-    for (size_t t = 0; t < T(); t++)
+    for (int t = 0; t < int(T()); t++)
     {
       fprintf(out,"%s %+3d %+2d %+2d %+2d  %9.17e %9.17e %9.17e %9.17e %9.17e %9.17e %9.17e %9.17e\n", prefix.c_str(), t,
               momentum[0], momentum[1], momentum[2],
